hotfix(workflows): inline evaluation-alignment workflow contract to unbreak deploy
The workflow was importing signal/query names and the state type from
`@domain/evaluations/alignment/workflow`, a subpath export intended to
avoid pulling `browser.ts` (and transitively `ServiceMap.Service`, which
mutates `Error.stackTraceLimit` and is forbidden in the Temporal workflow
sandbox) into the workflow bundle.

`@temporalio/worker`'s internal webpack bundler fails to resolve that
subpath at deploy time in the container image, crashing `Worker.create`
with `Module not found: Can't resolve '@domain/evaluations/alignment/workflow'`
and taking the ECS task with it. The deploy loops until timeout.

Sandbox-safe fix: inline the two strings and the narrow state type
directly into the workflow file. They are plain values with no runtime
dependencies, so there is nothing to bundle. `apps/web` and `apps/workers`
continue to import the same symbols from the `@domain/evaluations` barrel
where the sandbox constraint does not apply.

Follow-up: once the deploy is green we should either bump the Temporal SDK
to a webpack that handles this subpath, ship `@domain/evaluations` with
built `.js` output so webpack resolves without `.ts` loaders, or delete
the now-orphaned `packages/domain/evaluations/src/alignment/workflow.ts`
and its `exports` map entry.


refactor(evaluations): move alignment workflow contract into constants

Addresses review feedback on #2810: the prior hotfix inlined three
symbols (signal name, query name, state type) into the workflow file,
which duplicated them against `packages/domain/evaluations/src/alignment/workflow.ts`
and created a silent-drift risk between the workflow and the other
consumers (apps/web polling, apps/workers auto-refresh).

Resolution: move the three symbols into `@domain/evaluations/constants`,
which the workflow already imports from for the debounce windows and which
we know resolves inside `@temporalio/worker`'s webpack bundler (the
`./constants` subpath was never broken). Delete the now-unused
`src/alignment/workflow.ts` and its `./alignment/workflow` exports-map
entry. apps/web and apps/workers continue to import from the
`@domain/evaluations` barrel, which now re-exports from `constants.ts`.

Result: single source of truth, no sandbox regression, no subpath
resolution risk.
