+++ packages/domain/evaluations/package.json
@@ -10,8 +10,7 @@
       "browser": "./src/browser.ts",
       "default": "./src/index.ts"
     },
-    "./constants": "./src/constants.ts",
-    "./alignment/workflow": "./src/alignment/workflow.ts"
+    "./constants": "./src/constants.ts"
   },
   "scripts": {
     "check": "biome check src",
